noip

+++ Problem/src/蓝桥/Practise/A.java
@@ -1,9 +1,10 @@
 package 蓝桥.Practise;
 
 import java.util.Scanner;
+
 //两仪剑法
 public class A {
-    static int gcd(int a, int b) {
+    static long gcd(long a, long b) {
         if (b == 0) {
             return a;
         }
@@ -14,9 +15,9 @@ public class A {
         Scanner scanner = new Scanner(System.in);
         int t = scanner.nextInt();
         while (t-- > 0) {
-            int n = scanner.nextInt();
-            int m = scanner.nextInt();
-            int res = n * m / gcd(Math.max(n, m), Math.min(n, m));
+            long n = scanner.nextLong();
+            long m = scanner.nextLong();
+            long res = n * m / gcd(Math.max(n, m), Math.min(n, m));
             System.out.println(res);
         }
     }
